add room database

+++ core/types/src/main/java/de/franziskaneumeister/recentflics/core/types/entities/MovieId.kt
@@ -1,4 +1,4 @@
 package de.franziskaneumeister.recentflics.core.types.entities
 
 
-public typealias MovieId = Long
+public typealias MovieId = Int
